refactor(backend): вынести генерацию документов из views.py в offer/services/
- Создана структура offer/services/ с разделением по типу документа:
  base.py (render_docx, make_docx_response), offer_doc.py, contract.py,
  torg12.py, bill.py, context.py, __init__.py
- views.py сокращён с ~512 до 95 строк — только вызовы сервисов
- Добавлены select_related/prefetch_related во все запросы OfferItems
- Обновлены пути mock-патчей в тестах (api.v1.offer.views.DocxTemplate ->
  offer.services.offer_doc/base.DocxTemplate)
