+++ backend/offer/services/__init__.py
@@ -1,0 +1,4 @@
+from .offer_doc import generate_offer_doc
+from .contract import generate_contract_items_doc, generate_contract_service_doc
+from .torg12 import generate_torg12
+from .bill import generate_bill_work, generate_bill_items
